refactor: address review feedback — trim comments, split files, drop hydration fallback
- Drop narrative comments on three index UI/SDK files per review.
- Split files to 1-export-per-file (codebase convention):
  - buildIndexableSelectOptions.ts → 4 sibling files
  - compute-unique-field-metadata-ids-from-flat-indexes.util.ts →
    3 sibling files
  - create-index.input.ts → 3 sibling files
  - indexManifestType.ts → split out IndexFieldManifest
  - SettingsObjectNewIndex.tsx → extract FormValues + schema to
    a sibling file
- useCreateOneIndexMetadataItem: drop the input-hydration fallback
  and the speculative "cache can lag a beat" comment. The mutation
  selection already requests indexFieldMetadataList, which the
  server @ResolveField populates from the just-recomputed cache
  inside the same request — the fallback was defensive code added
  before the mutation selection itself was widened to include the
  list, and is now redundant. If the blank-row symptom reappears
  we'll investigate the cache pipeline rather than patch around it.
