+++ packages/twenty-front/src/modules/object-metadata/hooks/useCreateOneIndexMetadataItem.ts
@@ -3,7 +3,6 @@ import { CombinedGraphQLErrors } from '@apollo/client/errors';
 import { t } from '@lingui/core/macro';
 import { isDefined } from 'twenty-shared/utils';
 import { CrudOperationType } from 'twenty-shared/types';
-import { v4 as uuidv4 } from 'uuid';
 
 import { useMetadataErrorHandler } from '@/metadata-error-handler/hooks/useMetadataErrorHandler';
 import { useUpdateMetadataStoreDraft } from '@/metadata-store/hooks/useUpdateMetadataStoreDraft';
@@ -47,31 +46,17 @@ export const useCreateOneIndexMetadataItem = () => {
         const { __typename, indexFieldMetadataList, ...indexData } =
           createdIndex;
 
-        // Hydrate indexFieldMetadatas from the input rather than the
-        // server's @ResolveField on indexFieldMetadataList: the field-
-        // resolver reads from the flat-entity cache which can lag a beat
-        // behind the just-run migration and return an empty list, leaving
-        // the row labelled blank until a full refresh.
-        const hydratedIndexFieldMetadatas: IndexFieldMetadataItem[] =
-          indexFieldMetadataList.length === input.fields.length
-            ? indexFieldMetadataList.map(
-                ({ __typename: _, ...rest }) => rest as IndexFieldMetadataItem,
-              )
-            : input.fields.map((field, order) => ({
-                id: uuidv4(),
-                fieldMetadataId: field.fieldMetadataId,
-                subFieldName: field.subFieldName ?? undefined,
-                order,
-                createdAt: createdIndex.createdAt,
-                updatedAt: createdIndex.updatedAt,
-              }));
+        const indexFieldMetadatas: IndexFieldMetadataItem[] =
+          indexFieldMetadataList.map(
+            ({ __typename: _, ...rest }) => rest as IndexFieldMetadataItem,
+          );
 
         addToDraft({
           key: 'indexMetadataItems',
           items: [
             {
               ...indexData,
-              indexFieldMetadatas: hydratedIndexFieldMetadatas,
+              indexFieldMetadatas,
               objectMetadataId: input.objectMetadataId,
             } as FlatIndexMetadataItem,
           ],
