+++ packages/twenty-front/src/modules/settings/data-model/indexes/forms/components/SettingsObjectIndexOptionsForm.tsx
@@ -4,7 +4,7 @@ import { useLingui } from '@lingui/react/macro';
 import { Controller, useFormContext } from 'react-hook-form';
 import { themeCssVariables } from 'twenty-ui/theme-constants';
 import { IndexType } from '~/generated-metadata/graphql';
-import { type SettingsObjectNewIndexFormValues } from '~/pages/settings/data-model/new-index/SettingsObjectNewIndex';
+import { type SettingsObjectNewIndexFormValues } from '~/pages/settings/data-model/new-index/SettingsObjectNewIndexFormValues';
 
 const StyledContent = styled.div`
   display: flex;
