+++ packages/twenty-front/src/modules/settings/data-model/indexes/utils/encodeIndexableOptionValue.ts
@@ -1,0 +1,10 @@
+import { isNonEmptyString } from '@sniptt/guards';
+import { INDEXABLE_OPTION_SEPARATOR } from '@/settings/data-model/indexes/utils/indexableOptionSeparator';
+
+export const encodeIndexableOptionValue = (
+  fieldMetadataId: string,
+  subFieldName: string | null,
+): string =>
+  isNonEmptyString(subFieldName)
+    ? `${fieldMetadataId}${INDEXABLE_OPTION_SEPARATOR}${subFieldName}`
+    : fieldMetadataId;
