+++ packages/twenty-server/src/engine/metadata-modules/index-metadata/dtos/create-index.input.ts
@@ -6,28 +6,13 @@ import {
   ArrayNotEmpty,
   IsArray,
   IsEnum,
-  IsOptional,
-  IsString,
   IsUUID,
   ValidateNested,
 } from 'class-validator';
 
 import { UUIDScalarType } from 'src/engine/api/graphql/workspace-schema-builder/graphql-types/scalars';
+import { CreateIndexFieldInput } from 'src/engine/metadata-modules/index-metadata/dtos/create-index-field.input';
 import { IndexType } from 'src/engine/metadata-modules/index-metadata/types/indexType.types';
-
-@InputType()
-export class CreateIndexFieldInput {
-  @IsUUID()
-  @Field(() => UUIDScalarType)
-  fieldMetadataId!: string;
-
-  // Composite sub-property name (e.g. 'addressCity'). Required for composite
-  // parents, must be absent for scalar/relation parents.
-  @IsOptional()
-  @IsString()
-  @Field({ nullable: true })
-  subFieldName?: string;
-}
 
 @InputType()
 export class CreateIndexInput {
@@ -50,14 +35,4 @@ export class CreateIndexInput {
 
   // indexWhereClause is not exposed: the validator only allows a hardcoded
   // allowlist, so a free-text field on the user-facing API would mislead.
-}
-
-@InputType()
-export class CreateOneIndexInput {
-  @Type(() => CreateIndexInput)
-  @ValidateNested()
-  @Field(() => CreateIndexInput, {
-    description: 'The custom index to create',
-  })
-  index!: CreateIndexInput;
 }
