+++ packages/twenty-shared/src/application/index.ts
@@ -36,7 +36,8 @@ export type {
   CommandMenuItemManifest,
   FrontComponentManifest,
 } from './frontComponentManifestType';
-export type { IndexFieldManifest, IndexManifest } from './indexManifestType';
+export type { IndexFieldManifest } from './indexFieldManifestType';
+export type { IndexManifest } from './indexManifestType';
 export type {
   LogicFunctionManifest,
   CronTriggerSettings,
