+++ packages/twenty-shared/src/application/indexFieldManifestType.ts
@@ -1,0 +1,6 @@
+import { type SyncableEntityOptions } from '@/application/syncableEntityOptionsType';
+
+export type IndexFieldManifest = SyncableEntityOptions & {
+  fieldUniversalIdentifier: string;
+  subFieldName?: string;
+};
